+++ index.js
@@ -5,35 +5,40 @@ require('shelljs/global');
 var commands = require('./commands.json');
 
 program
+    .version('0.3.0')
     .option('-e, --everything', 'Purge all Docker containers, images and volumes; used or unused')
     .option('-i, --images', 'Purge all docker images with no associated containers')
     .option('-c, --containers', 'Purge all stopped Docker containers')
-    .option('-v, --volumes', 'Purge all unused Docker volumes')
+    .option('-x, --volumes', 'Purge all unused Docker volumes')
     .option('-a, --all', 'Append to other options to delete used images, containers or volumes as well. When deleting images or volumes this will result in all containers being deleted')
     .parse(process.argv);
 
 if(!which('docker')) {
-    console.log('In order to use this script, you must have Docker installed. See https://docs.docker.com/engine/installation/.');
+    console.log('In order to use this tool, you must have Docker installed. See https://docs.docker.com/engine/installation/.');
     exit(1);
 }
 
-if (program.containers || program.c) {
+checkVersion();
+
+if (!(program.containers || program.images || program.volumes || program.everything)) {
+    console.log('Please enter a valid option. Use --help for more information');
+    exit(1);
+}
+
+if (program.containers) {
     processInput(commands.containers);
 }
-if (program.images || program.i) {
+if (program.images) {
     processInput(commands.images);
 }
-if (program.volumes || program.v) {
+if (program.volumes) {
     processInput(commands.volumes);
 }
-if (program.everything || program.e) {
+if (program.everything) {
     runCommand(commands.containers, true);
     runCommand(commands.images);
     runCommand(commands.volumes);
     console.log('Everything purged');
-}
-else {
-    console.log('Please enter a valid option. Use --help for more information');
 }
 
 function processInput(command) {
@@ -56,5 +61,16 @@ function runCommand(type, all = false) {
     }
     else {
         exec(type.command, {silent: true});
+    }
+}
+
+function checkVersion() {
+    var dockerVersion = exec('docker --version', {silent:true}).stdout.split(',')[0];
+    var version = dockerVersion.split(' ').slice(-1).pop()
+    var versionSplit = version.split('.');
+    var versionShort = versionSplit[0] + '.' + versionSplit[1];
+    if (parseFloat(versionShort) < 1.13) {
+        console.log('You need Docker version 1.13.0 or above to use this tool.');
+        exit(1);
     }
 }