+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) 
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [0.3.0] - 2017-01-28
+### Changed
+- Added fix for check when no valid option is entered.
+- Added check for Docker version
+
+## [0.2.0] - 2017-01-28
+### Changed
+- Added check for when no valid option is entered
 
 ## [0.1.0] - 2017-01-28
 ### Added